requirements.txt
fastapi
uvicorn 
joblib
pandas
markdown2
slowapi
jinja2
sentence_transformers
faiss-cpu